requirements.txt
pandas
numpy
python-dotenv
pytz
requests
metaapi-cloud-sdk
websockets
asyncio
